+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Started changelog
 - Dynamic `<title>` in html
 - Rebrand from `Stencil` to `stencil.nu` in frontend
+- Added styling to EditSetView and made sure it and AddSetView is consistent with each other designwise
 ### Fixed
 - Frontend connects to backend in dev mode again
 - Favicon is now persistent after build
+- AddSetView now starts at the top when content is too tall